refactor: better structured routes/modules
