+++ src/lib/Stores/mockUser.ts
@@ -1,0 +1,8 @@
+import { writable } from 'svelte/store'
+
+export const userStore = writable({
+	username: 'johndoe',
+	password: 'abce',
+	name: 'John Doe',
+	admin: false
+})
