--- src/routes/account/_mockAccounts.ts
@@ -1,6 +1,0 @@
-export const user = {
-	username: 'johndoe',
-	password: 'abce',
-	name: 'John Doe',
-	admin: false
-};
