+++ src/routes/inloggen/index.svelte
@@ -2,9 +2,6 @@
 	import type { Load } from '@sveltejs/kit';
 
 	export const load: Load = ({ session, page }) => {
-		const { query } = page;
-		const redirectPage = query.get('page');
-		const logout = Boolean(query.get('logout'))
 
 		if (session.authenticated)
 			return {
@@ -13,32 +10,36 @@
 			};
 
 		return {
-			props: {
-				redirectPage
-			}
 		};
 	};
 </script>
 
 <script>
-	import { goto } from '$app/navigation';
 	import LoginForm from '$lib/account/LoginForm.svelte';
-	import { session } from '$app/stores';
+	import { goto } from '$app/navigation';
+	import { session as sessionStore, page } from '$app/stores';
+	import { messageStore } from '$lib/Stores/message'
 
-	const sessionInfo = $session
+	const { query } = $page
+	const session = $sessionStore
 
 	export let redirectPage;
-	export let error;
+
+	if(query.has('page')) {
+		redirectPage = query.get('page')
+		messageStore.set('Je moet eerst inloggen om deze pagina te zien')
+	}
 
 	function redirectToProfile(event) {
 		// The session needs to be written (only once) due to a Svelte Bug. `goto()` doens't give the cookie on redirects
-		session.set({ authenticated: true, user: event.detail.user });
+		sessionStore.set({ authenticated: true, user: event.detail.user });
 		if (redirectPage) goto(redirectPage);
 		goto('/account');
 	}
 
 	function handleError(event) {
-		error = event.detail.message;
+		let error = event.detail.message;
+		messageStore.set(error)
 	}
 </script>
 
@@ -46,11 +47,3 @@
 	<h1>Inloggen bij Jungle Gym</h1>
 	<LoginForm on:success={redirectToProfile} on:error={handleError} on:failure={handleError} />
 </main>
-
-{#if error}
-	<p>{error}</p>
-{/if}
-
-{#if redirectPage}
-	<p>Please login first to view this page</p>
-{/if}
